Move extension:dumpactive to own sub package
At some point we may want to split these off
