+++ Command/Extension/.gitattributes
@@ -1,0 +1,2 @@
+/Tests/ export-ignore
+/.gitattributes export-ignore
